app/FakeNewsDetectorAPI/requirements.txt
# django==4.2.3
# djangorestframework==3.14.0
# django-filter==24.2
# django-cors-headers==4.4.0
# psycopg2==2.9.9
# scikit-learn==1.5.0
# requests==2.32.3

django
djangorestframework
django-filter
django-cors-headers
psycopg2
scikit-learn
requests
bs4
numpy
pandas
matplotlib
seaborn
scikit-learn
wordcloud
nltk
mahotas
joblib
os
cv2
werkzeug










